// src/app/api_codegen/model/createCompanyDto.ts
/**
 * Invicto API
 * REST API Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CreateDepartmentDto } from './createDepartmentDto';

export interface CreateCompanyDto { 
    name: string;
    address: string;
    city: string;
    postalCode: string;
    country: string;
    costCenter?: CreateDepartmentDto;
    departments?: Array<CreateDepartmentDto>;
}